gateway/requirements.txt
flask==2.0.1
requests==2.26.0
prometheus-client==0.11.0
flask-healthz==0.0.3
werkzeug==2.0.3
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-exporter-otlp==1.20.0 